made a new file for Observable.flatMap
