+++ Appful/Observable.swift
@@ -69,36 +69,7 @@ class Observable<A> {
     func onComplete(handler: () -> Void){
         completionHandler = handler
     }
-    
-    //
-    // Composition API
-    //
-    
-    func flatMap<B>(f: A -> Observable<B>) -> Observable<B>{
-        return Observable<B>.create{ observer in
-            return self
-                .onSuccess{ (x: A) in
-                    let observable: Observable<B> = f(x)
-                    observable
-                        .onSuccess{ (y: B) in
-                            observer.success = y
-                        }
-                        .onError{ error in
-                            observer.error = error
-                        }
-                        .onComplete{
-                           observer.complete = true
-                        }
-                }
-                .onError{ error in
-                    observer.error = error
-                }
-                .onComplete{
-                }
-        }
-    }
 }
-
